first release commit
This is the code that passed the app certification kit tests on my
machine and has been submitted to the store for certification. Things
got really finicky with my ads. I had them all setup for the various
views and was able to see them, then all of the sudden no matter what I
tried they didn't show up any more. I'm hoping it's a problem on the
Microsoft side that will be resolved when the app is actually on the
market.